requirements.txt
aiogram==2.10.1
aiohttp==3.7.1
aioredis==1.3.1
python-dotenv
asyncpg
sqlalchemy
gino
